+++ components/navbar/index.ts
@@ -1,0 +1,6 @@
+export { default as Navbar } from "./Navbar.vue";
+export { default as NavbarDesktop } from "./NavbarDesktop.vue";
+export { default as NavbarMobile } from "./NavbarMobile.vue";
+export { default as NavbarSettingsDialog } from "./NavbarSettingsDialog.vue";
+export type { NavbarSection, NavigationItem } from "./navbar.types";
+