+++ components/skills/SkillsCard.vue
@@ -29,3 +29,4 @@ defineProps<Props>();
   </BentoGridItem>
 </template>
 
+