+++ components/skills/SkillsCarouselNavigation.vue
@@ -38,3 +38,4 @@ const emit = defineEmits<Emits>();
   </div>
 </template>
 
+