+++ components/skills/SkillsTabs.vue
@@ -53,3 +53,4 @@ const updateActiveTab = (tab: string) => {
   </div>
 </template>
 
+